add subscription options in user package
