+++ core/templates/site/user/subscriptions.gohtml
@@ -1,9 +1,8 @@
 {{ template "head" $ }}
 <h2>Your Subscriptions</h2>
-{{ if .Subs }}
 <table>
 <tr><th>Pattern</th><th>Method</th><th></th></tr>
-{{ range .Subs }}
+{{ range .UserSubscriptions }}
 <tr>
 <td>{{ .Pattern }}</td>
 <td>{{ .Method }}</td>
@@ -12,14 +11,13 @@
     {{ csrfField }}
     <input type="hidden" name="id" value="{{ .ID }}">
     <input type="submit" name="task" value="Delete">
-</form>
+    </form>
 </td>
 </tr>
+{{ else }}
+<tr><td colspan="3">No subscriptions</td></tr>
 {{ end }}
 </table>
-{{ else }}
-<p>No subscriptions</p>
-{{ end }}
 
 <h3>Update Subscriptions</h3>
 <form method="post" action="/usr/subscriptions/update">
@@ -27,8 +25,8 @@
     {{ range .Options }}
     <div>
         {{ .Name }}
-        <label><input type="checkbox" name="{{ .Path }}_internal" {{ if index $.SubMap (printf "%s|internal" .Pattern) }}checked{{ end }}>Internal</label>
-        <label><input type="checkbox" name="{{ .Path }}_email" {{ if index $.SubMap (printf "%s|email" .Pattern) }}checked{{ end }}>Email</label>
+        <label><input type="checkbox" name="{{ .Path }}_internal" {{ if $.Subscribed .Pattern "internal" }}checked{{ end }}>Internal</label>
+        <label><input type="checkbox" name="{{ .Path }}_email" {{ if $.Subscribed .Pattern "email" }}checked{{ end }}>Email</label>
     </div>
     {{ end }}
     <input type="submit" name="task" value="Update">
